Move new tables to components directory
