+++ src/components/AssociatedIndexVariantsTable.js
